pyproject.toml
[project]
name = "funhub-api"
version = "0.1.0"
description = "FunHub backend FastAPI service"
authors = [{ name = "FunHub" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "supabase>=2.4.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    "pyjwt>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.uvicorn]
reload = true
factory = true
app = "app.main:create_app"
